backend/requirements.txt
alembic==1.17.0
annotated-types==0.7.0
anyio==4.11.0
bcrypt==4.0.1
cffi==2.0.0
click==8.3.0
colorama==0.4.6
cryptography==46.0.2
fastapi==0.119.0
greenlet==3.2.4
h11==0.16.0
idna==3.11
isort==7.0.0
jose==1.0.0
Mako==1.3.10
MarkupSafe==3.0.3
passlib==1.7.4
psycopg2-binary==2.9.10
pycparser==2.23
pydantic==2.12.2
pydantic_core==2.41.4
PyJWT==2.10.1
python-dotenv==1.1.1
sniffio==1.3.1
SQLAlchemy==2.0.44
starlette==0.48.0
types-passlib==1.7.7.20250602
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.37.0
